+++ packages/terra-data-grid/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Fixed
+  * Fixed tab focus management when next element is in a hidden container.
+  * Fixed tab focus maanagement to wrap to the first focusable element when the data grid is the last focusable element.
+
 ## 0.6.0 - (September 14, 2023)
 
 * Breaking Changes